+++ Magic/CHANGELOG.md
@@ -5,6 +5,8 @@
 ## Fixes
  
  - Potion effect attributes changed to be 1-based
+ - Enchantment level attributes will show the maximum of all armor and items
+   Note that it's still not super smart, holding an armor piece will count
 
 ## Additions